added (static) zoom
Added --zoom option to allow a static zoom factor to be applied when
generating images of snapshots. (Eventual plan is to allow this to be
done as a function of simulation time.)

+++ SimulationTools/zeno/movieframes.py
@@ -14,7 +14,7 @@
 import sys,re,argparse,os,glob
 
 VERSION='0.1a'
-VERSIONDATE='06 July 2012'
+VERSIONDATE='19 July 2012'
 
 # set up command line arguments
 parser=argparse.ArgumentParser(description='Generate pgm images of subsets of a zeno N-body snapshot. Assumes first file has BodyType information in it.')
@@ -23,6 +23,7 @@ parser.add_argument('-DM',action='store_true',default=False,help='Select dark ma
 parser.add_argument('-sdisk',action='store_true',default=False,help='Select stellar disk particles')
 parser.add_argument('-sbulge',action='store_true',default=False,help='Select stellar bulge particles')
 parser.add_argument('-gdisk',action='store_true',default=False,help='Select gas disk particles')
+parser.add_argument('--zoom',action='store',default=False,help='Zoom in (<1) or out (>1) by the factor specified')
 parser.add_argument('--viewfile=',action='store',default=False,help='Specify a view file to rotate the snapshots prior to generating the images. [NOT YET IMPLEMENTED].')
 parser.add_argument('--version',action='version',version='%(prog)s '+VERSION+' ('+VERSIONDATE+')')
 args=parser.parse_args()
@@ -73,19 +74,27 @@ snaps=args.snaps
 for snap in snaps:
   # generate an output file name
   oname=snap.split('.dat')[0]+suffix+'.pgm'
-
-  # make sure our temp files don't already exist
-  if os.path.exists('tempsnap.dat'):
-    os.remove('tempsnap.dat')
-  if os.path.exists('tmpimg.dat'):
-    os.remove('tmpimg.dat')
-
-  # now generate the snapshots
-  os.system('csf '+snaps[0]+' ./tempsnap.dat')
-  os.system('csf '+snap+' ./tempsnap.dat append=t')
-  os.system('snapcollect tempsnap.dat - | snapsift - - sieve="'+sieve+'" | snapset - - produce=BodyType,Uinternal,Mass type=0x60 uint=0.00001 m=0.00001 z=0 | sphcode_u - tmpimg.dat tstop=0 outputs=Position,SmoothLength')
-  os.system('snapset tmpimg.dat - aux=0.0001 produce=Aux | snapsmooth - '+oname+' value=bright threedim=false')
-
-  # delete our temp files
-  os.remove('tmpimg.dat')
-  os.remove('tempsnap.dat')
+  # is there already an image for this filename?
+  if os.path.exists(oname):
+    sys.stderr.write(oname+' already exists, skipping.\n')
+  else: 
+    # make sure our temp files don't already exist
+    if os.path.exists('tempsnap.dat'):
+      os.remove('tempsnap.dat')
+    if os.path.exists('tmpimg.dat'):
+      os.remove('tmpimg.dat')
+  
+    # now generate the snapshots
+    os.system('csf '+snaps[0]+' ./tempsnap.dat')
+    os.system('csf '+snap+' ./tempsnap.dat append=t')
+    if args.zoom:
+      os.system('snapcollect tempsnap.dat - | snapsift - - sieve="'+sieve+'" | snapset - - x="x/'+str(args.zoom)+'" y="y/'+str(args.zoom)+'" produce=BodyType,Uinternal,Mass type=0x60 uint=0.00001 m=0.00001 z=0 | sphcode_u - tmpimg.dat tstop=0 outputs=Position,SmoothLength')
+    else: 
+      os.system('snapcollect tempsnap.dat - | snapsift - - sieve="'+sieve+'" | snapset - - produce=BodyType,Uinternal,Mass type=0x60 uint=0.00001 m=0.00001 z=0 | sphcode_u - tmpimg.dat tstop=0 outputs=Position,SmoothLength')
+    os.system('snapset tmpimg.dat - aux=0.0001 produce=Aux | snapsmooth - '+oname+' value=bright threedim=false zval=2.0')
+  
+    # delete our temp files
+    if os.path.exists('tmpimg.dat'):
+      os.remove('tmpimg.dat')
+    if os.path.exists('tempsnap.dat'):
+      os.remove('tempsnap.dat')
